+++ pkg/agent/collectors/metrics/configmap.go
@@ -82,13 +82,23 @@ func (a *Service) GenerateAlloyMonitoringConfigMapData(ctx context.Context, curr
 	}
 
 	data := struct {
-		AlloyConfig       string
-		PriorityClassName string
-		Replicas          int
+		AlloyConfig                    string
+		PriorityClassName              string
+		Replicas                       int
+		KEDAAuthenticationEnabled      bool
+		AlloySecretName                string
+		MimirRemoteWriteAPIUsernameKey string
+		MimirRemoteWriteAPIPasswordKey string
+		MimirQueryAPIURLKey            string
 	}{
-		AlloyConfig:       alloyConfig,
-		PriorityClassName: common.PriorityClassName,
-		Replicas:          shards,
+		AlloyConfig:                    alloyConfig,
+		PriorityClassName:              common.PriorityClassName,
+		Replicas:                       shards,
+		KEDAAuthenticationEnabled:      a.Config.Monitoring.IsKEDAAuthenticationEnabled(cluster),
+		AlloySecretName:                apps.AlloyMetricsAppName,
+		MimirRemoteWriteAPIUsernameKey: mimirRemoteWriteAPIUsernameKey,
+		MimirRemoteWriteAPIPasswordKey: mimirRemoteWriteAPIPasswordKey,
+		MimirQueryAPIURLKey:            common.MimirQueryAPIURLKey,
 	}
 
 	var values bytes.Buffer